requirements.txt
python-telegram-bot==20.7
pymongo==4.10.1
dnspython==2.6.1
